utils: Move open_dir_noxdev here
Prep for using it elsewhere.
